+++ CHANGELOG.md
@@ -3,7 +3,20 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
+
+## [26.0.1] - 2026-02-12
+
+### Added
+
+- v0.1.28 STATUS tests
+- v0.1.28 RESPONSE tests
+
+### Fixed
+
+- Wrong namespace used
+
+
 
 ## [25.0.0] - 2025-12-13